refactor: extract browser execution functions for skill wrapper reuse
